+++ src/App.scss
@@ -13,13 +13,3 @@
   padding: 20px;
   border-left: 1px solid lightgray;
 }
-
-.movies {
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(250px, 1fr));
-  gap: 20px;
-}
-
-.card-image img {
-  object-fit: contain;
-}
