+++ src/components/MovieCard/MovieCard.scss
@@ -1,1 +1,4 @@
-// 
+//
+.card-image img {
+  object-fit: contain;
+}
