+++ src/components/MoviesList/MoviesList.scss
@@ -1,1 +1,6 @@
-// 
+//
+.movies {
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(250px, 1fr));
+  gap: 20px;
+}
